perf: optimize CSS selectors (#341)
Simple list selectors seem to be performing better. From initial testing
in the complex app it shaved off around 50ms of "elapsed time" in
selector stats in performance profiler. This was tested via sources
override in chrome devtools which is quite buggy, that's why I want to
release updated version, install it in our app and test in preview.